requirements.txt
altair==5.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.9.2
nest-asyncio==1.5.9
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.30.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
wcwidth==0.2.13
wordcloud==1.9.3
zipp==3.17.0
